Draw the evolution of the decision boundary of NN
Draw the evolution of the decision boundary of a Neural Network through the training process. I have a default data builder and a default Neural Network, but you can change those and the code will still (hopefully**) work. Only constraints are that the input must be a 2D matrix,
but you can change build_2D_data()  function, to create other inputs.